add delete active (foodItems, category, shelfLife, restaurant)

+++ SpringFoodPlacesWeb/src/main/java/com/tuantran/repository/impl/CategoryFoodRepositoryImpl.java
@@ -5,7 +5,9 @@
 package com.tuantran.repository.impl;
 
 import com.tuantran.pojo.CategoriesFood;
+import com.tuantran.pojo.Fooditems;
 import com.tuantran.repository.CategoryFoodRepository;
+import com.tuantran.service.FoodItemsService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -35,9 +37,13 @@ public class CategoryFoodRepositoryImpl implements CategoryFoodRepository {
 
     @Autowired
     private LocalSessionFactoryBean factory;
+    
     @Autowired
     private Environment environment;
-
+    
+    @Autowired
+    private FoodItemsService foodItemsService;
+    
 //    @Override
 //    public List<Object[]> getCategoriesFood(Map<String, String> params) {
 //        Session session = this.factory.getObject().getCurrentSession();
@@ -139,7 +145,18 @@ public class CategoryFoodRepositoryImpl implements CategoryFoodRepository {
         Session session = this.factory.getObject().getCurrentSession();
         CategoriesFood cate = this.getCategoryById(id);
         try {
-            session.delete(cate);
+            if (cate.getActive().equals(Boolean.TRUE)) {
+                cate.setActive(Boolean.FALSE);
+                session.update(cate);
+                List<Fooditems> foodItem_list = this.foodItemsService.getFoodItemsByCategoryId(id);
+                
+                for (Fooditems food : foodItem_list) {
+                    this.foodItemsService.delFoodItem(food.getFoodId());
+                }
+                
+            } else {
+                session.delete(cate);
+            }
             return true;
         } catch (HibernateException ex) {
             ex.printStackTrace();
